+++ openapi-specs/package.json
@@ -1,6 +1,8 @@
 {
   "name": "@saflib/openapi-specs",
   "description": "Shared dependencies for generating OpenAPI specs in SAF projects",
+  "type": "module",
+  "main": "./index.ts",
   "dependencies": {
     "@redocly/cli": "^1.0.0",
     "openapi-typescript": "^7.0.0",
